back-end/requirements.txt
PyPDF2
fastapi
uvicorn
google-genai
google
matplotlib
python-dotenv
pandas
openpyxl